Replace np.bool with bool in GraphCodeBERT
`np.bool` was a deprecated alias. It is used in GraphCodeBERT/codesearch/run.py in the TextDataset. This TextDataset is used by CodeCSE to prepare the input features. So we replace `np.bool` with `bool` as a quick fix.

The original error message from numpy:
```
`np.bool` was a deprecated alias for the builtin `bool`. To avoid this error in existing code, use `bool` by itself. Doing this will not modify any behavior and is safe. If you specifically wanted the numpy scalar type, use `np.bool_` here.
The aliases was originally deprecated in NumPy 1.20; for more details and guidance see the original release note at:
    https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations
```